# openapi/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CancelContractRequest:
    properties:
      note:
        type: string
    type: object
...
  handlers.RejectContractRequest:
    properties:
      reason:
        type: string
    type: object
...
  models.ContractLedgerEntryResponse:
    properties:
      amount:
        type: number
      contract_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      entry_date:
        type: string
      entry_type:
        type: string
      id:
        type: integer
      payment_id:
        type: integer
      updated_at:
        type: string
    type: object
  models.ContractResponse:
    properties:
      amount:
        type: number
      applicant_credit_score:
        type: integer
      applicant_identity:
        type: string
      applicant_name:
        type: string
      applicant_phone:
        type: string
      applicant_user_id:
        type: integer
      approved_at:
        type: string
      balance:
        type: number
      cancellation_notes:
        type: string
      contract_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      down_payment:
        type: number
      financing_type:
        type: string
      id:
        type: integer
      ledger_entries:
        items:
          $ref: '#/definitions/models.ContractLedgerEntryResponse'
        type: array
      lot_address:
        type: string
      lot_area:
        type: number
      lot_id:
        type: integer
      lot_length:
        type: number
      lot_name:
        type: string
      lot_override_price:
        type: number
      lot_price:
        type: number
      lot_width:
        type: number
      note:
        type: string
      payment_schedule:
        items:
          $ref: '#/definitions/models.PaymentResponse'
        type: array
      payment_term:
        type: integer
      project_address:
        type: string
      project_id:
        type: integer
      project_name:
        type: string
      rejection_reason:
        type: string
      reserve_amount:
        type: number
      status:
        type: string
      total_interest_collected:
        type: number
      total_paid:
        type: number
      updated_at:
        type: string
    type: object
...
  models.PaymentResponse:
    properties:
      amount:
        type: number
      applicant_identity:
        type: string
      applicant_name:
        type: string
      applicant_phone:
        type: string
      approved_at:
        type: string
      contract_id:
        type: integer
      contract_status:
        description: Contract details
        type: string
      description:
        type: string
      due_date:
        type: string
      has_receipt:
        type: boolean
      id:
        type: integer
      interest_amount:
        type: number
      lot_address:
        type: string
      lot_dimensions:
        type: string
      lot_name:
        type: string
      overdue_days:
        type: integer
      paid_amount:
        type: number
      payment_date:
        type: string
      payment_type:
        type: string
      project_address:
        type: string
      project_name:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for Fintera Real Estate Management System
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fintera API
  version: "1.0"
paths:
...
  /contracts:
    get:
      consumes:
      - application/json
      description: Get a paginated list of contracts for the current user (or all
        for admin)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Search term
        in: query
        name: search_term
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Contracts
      tags:
      - Contracts
...
  /projects/{project_id}/lots/{lot_id}/contracts:
    post:
      consumes:
      - multipart/form-data
      description: Create a new contract with optional new user and documents
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      - description: Contract Data (JSON)
        in: formData
        name: contract
        required: true
        type: string
      - description: User Data (JSON)
        in: formData
        name: user
        type: string
      - description: Documents
        in: formData
        name: documents
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create Contract
      tags:
      - Contracts
  /projects/{project_id}/lots/{lot_id}/contracts/{contract_id}:
    get:
      consumes:
      - application/json
      description: Get a contract by ID
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ContractResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Contract
      tags:
      - Contracts
    patch:
      consumes:
      - application/json
      description: Update a contract
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      - description: Contract Data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Contract
      tags:
      - Contracts
  /projects/{project_id}/lots/{lot_id}/contracts/{contract_id}/approve:
    post:
      consumes:
      - application/json
      description: Approve a pending contract (Admin)
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ContractResponse'
      security:
      - BearerAuth: []
      summary: Approve Contract
      tags:
      - Contracts
  /projects/{project_id}/lots/{lot_id}/contracts/{contract_id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel an active contract (Admin)
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      - description: Note
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CancelContractRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ContractResponse'
      security:
      - BearerAuth: []
      summary: Cancel Contract
      tags:
      - Contracts
  /projects/{project_id}/lots/{lot_id}/contracts/{contract_id}/capital_repayment:
    post:
      consumes:
      - application/json
      description: Apply capital repayment to contract (Admin)
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Capital Repayment
      tags:
      - Contracts
  /projects/{project_id}/lots/{lot_id}/contracts/{contract_id}/ledger:
    get:
      consumes:
      - application/json
      description: Get ledger entries for a contract
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Contract Ledger
      tags:
      - Contracts
...
  /projects/{project_id}/lots/{lot_id}/contracts/{contract_id}/reject:
    post:
      consumes:
      - application/json
      description: Reject a pending contract (Admin)
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      - description: Reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RejectContractRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ContractResponse'
      security:
      - BearerAuth: []
      summary: Reject Contract
      tags:
      - Contracts
  /projects/{project_id}/lots/{lot_id}/contracts/{contract_id}/reopen:
    post:
      consumes:
      - application/json
      description: Reopen a cancelled contract (Admin)
      parameters:
      - description: Contract ID
        in: path
        name: contract_id
        required: true
        type: integer
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reopen Contract
      tags:
      - Contracts
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"